fix sample commands, rename, text fixes for langserve
